requirement.txt
matplotlib==3.1.3
numpy==1.18.1
pandas==1.0.2
algorithm==1.2.0
scikit_learn==0.23.2
cvxpy==1.0.28
detector==3.0
PyYAML==5.3.1
utils==1.0.1
scipy==1.4.1
tqdm==4.43.0
pylbfgs==0.2.0.13
